fix for loading petdata - tnx pompinis & GodKratos
git-svn-id: http://svn.l2jserver.com/trunk@2377 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/net/sf/l2j/gameserver/model/L2PetDataTable.java
@@ -65,7 +65,7 @@ public class L2PetDataTable
     			L2PetData petData = new L2PetData();
     			petData.setPetID(petId);
                 petData.setPetLevel(petLevel);
-                petData.setPetMaxExp(rset.getInt("expMax"));
+                petData.setPetMaxExp(rset.getLong("expMax"));
                 petData.setPetMaxHP(rset.getInt("hpMax"));
                 petData.setPetMaxMP(rset.getInt("mpMax"));
                 petData.setPetPAtk( rset.getInt("patk") );
